messenger/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
starlette==0.36.3
websockets==12.0 